Use internal type for country storage
This lets us use the same type in our revolvers for dealing with
countries.

If `avoidOptionals` is turned on, then a resolver for each type needs
to be defined for the `Resolver` type generated by graphql-code-gen.
This is fairly ugly, as all the types used in GraphQL (except for
Country), are 1:1 mappings between their TS Type and GraphQL.

Have switched to graphql-typescript-simple-resolvers which avoids this
issue